Update patch set 1
Patch Set 1:

This is the same linking strategy I used for the monitoring section. The URL hierarchy on the verify job is different than it is on the official latest which looks like it's causing the links to break.   I would suggest we merge this, check the links in prod, and then fix them if needed.  I think it will work.

Patch-set: 1
Reviewer: Gerrit User 3999 <3999@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0